requirements.txt
Django==2.2.3
django-appconf==1.0.3
django-image-cropping==1.2.0
easy-thumbnails==2.6
Pillow==6.1.0
psycopg2-binary==2.8.3
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
